+++ packages/aiinfo/node/index.ts
@@ -4,6 +4,7 @@ import { canvascoursetranslation } from "./components/canvascoursetranslation";
 import { canvasdiscussionsummaries } from "./components/canvasdiscussionsummaries";
 import { canvasgradingassistance } from "./components/canvasgradingassistance";
 import { canvasinboxtranslation } from "./components/canvasinboxtranslation";
+import { conversionalignment } from "./components/conversionalignment";
 import { discussioninsights } from "./components/discussioninsights";
 import { igniteagent } from "./components/igniteagent";
 import { portfolios } from "./components/portfolios";
@@ -41,6 +42,7 @@ const AiInfo: AiInfoProps = {
 	canvasdiscussionsummaries,
 	canvasgradingassistance,
 	canvasinboxtranslation,
+	conversionalignment,
 	discussioninsights,
 	igniteagent,
 	portfolios,
@@ -68,6 +70,7 @@ export {
 	canvasdiscussionsummaries,
 	canvasgradingassistance,
 	canvasinboxtranslation,
+	conversionalignment,
 	discussioninsights,
 	igniteagent,
 	portfolios,